Added ClassModels folder Removed old v1 asset loading on v2 Moved AssetsJsonWrapper to ClassModels
