+++ Assets/Scripts/ClassModels/ClassModels.asmdef
@@ -1,0 +1,3 @@
+﻿{
+  "name": "ClassModels"
+}
