+++ Assets/Scripts/ClassModels/ClassModels.asmdef.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: d8bd096b2dbb495485e42270be514f96
+timeCreated: 1678737154
