+++ Assets/Scripts/System/Systems.asmdef
@@ -13,7 +13,8 @@
         "GUID:e0946e506e3794c42a0a4080d32eea3a",
         "GUID:75469ad4d38634e559750d17036d5f7c",
         "GUID:dfd4dd069f777a943b33ee54f1a6a649",
-        "GUID:69f072ba8fc33de469fe24734c3f64fa"
+        "GUID:69f072ba8fc33de469fe24734c3f64fa",
+        "GUID:d8bd096b2dbb495485e42270be514f96"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
